+++ bauh/commons/system.py
@@ -106,7 +106,7 @@ class ProcessHandler:
         if self.watcher:
             self.watcher.print(msg)
 
-    def handle(self, process: SystemProcess, error_output: StringIO = None) -> bool:
+    def handle(self, process: SystemProcess, error_output: StringIO = None, output_handler=None) -> bool:
         self._notify_watcher(' '.join(process.subproc.args) + '\n')
 
         already_succeeded = False
@@ -116,6 +116,9 @@ class ProcessHandler:
                 line = output.decode().strip()
                 if line:
                     self._notify_watcher(line)
+
+                    if output_handler:
+                        output_handler(line)
 
                     if process.success_phrases and [p in line for p in process.success_phrases]:
                         already_succeeded = True
@@ -131,6 +134,8 @@ class ProcessHandler:
                 line = output.decode().strip()
                 if line:
                     self._notify_watcher(line)
+
+                    output_handler(line)
 
                     if error_output is not None:
                         error_output.write(line)