+++ bauh/view/core/config.py
@@ -19,8 +19,7 @@ def read_config(update_file: bool = False) -> dict:
         'locale': None,
         'updates': {
             'check_interval': 30,
-            'sort_packages': True,
-            "pre_dependency_checking": True
+            'sort_packages': True
         },
         'system': {
           'notifications': True,